+++ tensorflow_similarity/augmenters/augmenter.py
@@ -15,6 +15,7 @@
 
 import abc
 from tensorflow_similarity.types import Tensor
+# from tensorflow import Tensor
 from typing import List
